requirements.txt
flask
gunicorn
numpy
pandas
twilio
unidecode
python-dotenv
Pillow  # Em vez de PIL
pytesseract
requests
werkzeug
flask-wtf
xlsxwriter
openpyxl
jsonify
pydub
pymupdf
matplotlib
plotly
psycopg2-binary
supabase


# flask==2.0.1
# gunicorn==20.1.0
# numpy==1.22.4
# pandas==1.4.2
# twilio==7.9.0
# unidecode==1.3.4
# python-dotenv==0.19.2